python_system/moderator_services/moderation_service/requirements.txt
# =============================================================================
# AdSphere Moderation Service - Requirements
# =============================================================================

# FastAPI + Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
uvloop>=0.19.0
httptools>=0.6.0
websockets>=12.0
python-multipart>=0.0.6

# Async + Concurrency
aiohttp>=3.9.0
aiofiles>=23.2.0
asyncio>=3.4.3

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ML/AI Core
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Video Processing
ffmpeg-python>=0.2.0

# NLP
spacy>=3.7.0
fasttext>=0.9.2
detoxify>=0.5.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Serialization
orjson>=3.9.0
protobuf>=4.25.0

# Monitoring
prometheus-client>=0.19.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=23.2.0
psutil>=5.9.0

